backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pymongo==4.10.1
python-jose==3.3.0
pydantic==2.10.3
pydantic[email]==2.10.3
pdfplumber==0.11.3
groq==0.9.0
bcrypt==4.2.1
python-dotenv==1.0.1
